introduce acme.sh service
I did not follow the default recepies of restarting nginx on
sertificates renewal, as it requires to pass /var/run/docker.sock to
the acme container, so it able to restart nginx container. The problem
is acme would've got the full control over the docker, which is not
ideal from the security point of view.

Instead, a crontab entry is added to the nginx container to reload
sertificaes weekly. From the up time (if it even matters), this is
even beter, because "nginx -s reload" compleates in no time. While a
full container reloading might take a second or two.

Signed-off-by: Yuri Volchkov <yuri@volch.org>